requirements.txt
python-telegram-bot==21.3
python-dotenv==1.0.1
httpx==0.28.1
google-genai==1.5.0
